+++ packages/yourloops/lib/share/api.ts
@@ -27,7 +27,6 @@
  */
 
 import bows from "bows";
-import _ from "lodash";
 
 import { UserInvitationStatus } from "../../models/generic";
 import { UserRoles } from "../../models/shoreline";
@@ -57,10 +56,10 @@ async function getDirectShares(session: Session): Promise<ShareUser[]> {
     const directShares = (await response.json()) as DirectShareAPI[];
     const shareUsers: ShareUser[] = [];
     if (Array.isArray(directShares)) {
-      for (let i=0; i<directShares.length; i++) {
+      for (let i = 0; i < directShares.length; i++) {
         const directShare = directShares[i];
         const directShareWith = directShare.patient ?? directShare.viewer;
-        if (_.isNil(directShareWith)) {
+        if (!directShareWith) {
           continue;
         }
 
@@ -74,6 +73,7 @@ async function getDirectShares(session: Session): Promise<ShareUser[]> {
             username: directShareWith.email,
             emails: [directShareWith.email],
             role: user.role === UserRoles.patient ? UserRoles.caregiver : UserRoles.patient,
+            idVerified: directShareWith.idVerified,
           },
         };
         shareUsers.push(shareUser);