+++ packages/yourloops/pages/patient/caregivers/table.tsx
@@ -29,7 +29,8 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 
-import { makeStyles, Theme } from "@material-ui/core/styles";
+import { makeStyles } from "@material-ui/core/styles";
+import Box from "@material-ui/core/Box";
 import Paper from "@material-ui/core/Paper";
 import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
@@ -43,6 +44,7 @@ import Tooltip from "@material-ui/core/Tooltip";
 import AccessTimeIcon from "@material-ui/icons/AccessTime";
 import PersonRemoveIcon from "../../../components/icons/PersonRemoveIcon";
 import IconActionButton from "../../../components/buttons/icon-action";
+import CertifiedProfessionalIcon from "../../../components/icons/certified-professional-icon";
 
 import { UserInvitationStatus } from "../../../models/generic";
 import { ShareUser } from "../../../lib/share";
@@ -56,26 +58,10 @@ export interface CaregiverTableProps {
 }
 
 const tableStyles = makeStyles(
-  (theme: Theme) => {
+  () => {
     return {
-      table: {
-        width: "100%",
-      },
-      tableRow: {
-        cursor: "pointer",
-      },
-      tableRowPending: {
-        cursor: "pointer",
-        backgroundColor: theme.palette.secondary.main,
-      },
       tableRowHeader: {
         textTransform: "uppercase",
-      },
-      tableCellHeader: {
-        fontSize: "14px",
-      },
-      tableCellHeaderPending: {
-        width: "24px",
       },
       tableCellActions: {
         textAlign: "right",
@@ -121,7 +107,7 @@ function CaregiverTable(props: CaregiverTableProps): JSX.Element {
   const [order, setOrder] = React.useState<SortDirection>(SortDirection.asc);
 
   const createSortHandler = (property: SortFields): (() => void) => {
-    return (/* event: React.MouseEvent */): void => {
+    return (): void => {
       if (orderBy === property) {
         setOrder(order === SortDirection.asc ? SortDirection.desc : SortDirection.asc);
       } else {
@@ -135,45 +121,42 @@ function CaregiverTable(props: CaregiverTableProps): JSX.Element {
     return compareUserShare(orderBy, order, a, b);
   });
 
-  const tableRows = props.userShares.map((us) => {
-    const userId = us.user.userid;
-    const handleClickDelete = () => {
-      props.onRemoveCaregiver(us);
-    };
-    const removeAction = t("modal-patient-remove-caregiver-remove");
+  const tableRows = props.userShares.map((userShare) => {
+    const userId = userShare.user.userid;
 
-    if (us.status === UserInvitationStatus.pending) {
-      return (
-        <TableRow key={userId} id={`patient-caregivers-table-row-${userId}`} data-userid={userId} data-status={us.status} className={classes.tableRowPending}>
-          <TableCell id={`patient-caregivers-table-row-${userId}-status`}>
-            <Tooltip title={t("pending-invitation") as string} aria-label={t("pending-invitation")} placement="bottom">
-              <AccessTimeIcon id={`patient-caregivers-table-row-pendingicon-${userId}`} />
-            </Tooltip>
-          </TableCell>
-          <TableCell id={`patient-caregivers-table-row-${userId}-lastname`}>-</TableCell>
-          <TableCell id={`patient-caregivers-table-row-${userId}-firstname`}>{us.user.username}</TableCell>
-          <TableCell id={`patient-caregivers-table-row-${userId}-actions`} className={classes.tableCellActions}>
-            <IconActionButton
-              icon={<PersonRemoveIcon />}
-              id={`patient-caregivers-table-row-${userId}-button-remove`}
-              onClick={handleClickDelete}
-              tooltip={removeAction}
-            />
-          </TableCell>
-        </TableRow>
-      );
-    }
     return (
-      <TableRow key={userId} id={`patient-caregivers-table-row-${userId}`} data-userid={userId} data-status={us.status}>
-        <TableCell id={`patient-caregivers-table-row-${userId}-status`} />
-        <TableCell id={`patient-caregivers-table-row-${userId}-lastname`}>{getUserLastName(us.user)}</TableCell>
-        <TableCell id={`patient-caregivers-table-row-${userId}-firstname`}>{getUserFirstName(us.user)}</TableCell>
+      <TableRow
+        key={userId}
+        id={`patient-caregivers-table-row-${userId}`}
+        data-userid={userId} data-status={userShare.status}
+      >
+        <TableCell id={`patient-caregivers-table-row-${userId}-status`}>
+          {userShare.status === UserInvitationStatus.pending &&
+            <Box display="flex">
+              <Tooltip title={t("pending-invitation") as string} aria-label={t("pending-invitation")} placement="bottom">
+                <AccessTimeIcon id={`patient-caregivers-table-row-pendingicon-${userId}`} />
+              </Tooltip>
+            </Box>
+          }
+        </TableCell>
+        <TableCell id={`patient-caregivers-table-row-${userId}-lastname`}>
+          {userShare.status === UserInvitationStatus.pending ? "-" : getUserLastName(userShare.user)}
+        </TableCell>
+        <TableCell id={`patient-caregivers-table-row-${userId}-firstname`}>
+          {userShare.status === UserInvitationStatus.pending ? "-" : getUserFirstName(userShare.user)}
+        </TableCell>
+        <TableCell id={`patient-caregivers-table-row-${userId}-email`}>
+          <Box display="flex">
+            {userShare.user.username}
+            {userShare.user.idVerified && <CertifiedProfessionalIcon />}
+          </Box>
+        </TableCell>
         <TableCell id={`patient-caregivers-table-row-${userId}-actions`} className={classes.tableCellActions}>
           <IconActionButton
             icon={<PersonRemoveIcon />}
             id={`patient-caregivers-table-row-${userId}-button-remove`}
-            onClick={handleClickDelete}
-            tooltip={removeAction}
+            onClick={() => props.onRemoveCaregiver(userShare)}
+            tooltip={t("modal-patient-remove-caregiver-remove")}
           />
         </TableCell>
       </TableRow>
@@ -182,27 +165,28 @@ function CaregiverTable(props: CaregiverTableProps): JSX.Element {
 
   return (
     <TableContainer component={Paper}>
-      <Table id="patient-caregivers-table" className={classes.table} aria-label={t("aria-table-list-caregivers")} stickyHeader>
+      <Table id="patient-caregivers-table" aria-label={t("aria-table-list-caregivers")} stickyHeader>
         <TableHead id="patient-caregivers-table-header">
           <TableRow id="patient-caregivers-table-header-row" className={classes.tableRowHeader}>
-            <TableCell
-              id="patient-caregivers-table-header-status"
-              className={`${classes.tableCellHeader} ${classes.tableCellHeaderPending}`}
-            />
-            <TableCell id="patient-caregivers-table-header-lastname" className={classes.tableCellHeader}>
+            <TableCell id="patient-caregivers-table-header-status" />
+            <TableCell id="patient-caregivers-table-header-lastname">
               <TableSortLabel active={orderBy === "lastname"} direction={order} onClick={createSortHandler(SortFields.lastname)}>
                 {t("lastname")}
               </TableSortLabel>
             </TableCell>
-            <TableCell id="patient-caregivers-table-header-firstname" className={classes.tableCellHeader}>
+            <TableCell id="patient-caregivers-table-header-firstname">
               <TableSortLabel
                 active={orderBy === "firstname"}
                 direction={order}
-                onClick={createSortHandler(SortFields.firstname)}>
+                onClick={createSortHandler(SortFields.firstname)}
+              >
                 {t("firstname")}
               </TableSortLabel>
             </TableCell>
-            <TableCell id="patient-caregivers-table-header-actions" className={classes.tableCellHeader} />
+            <TableCell id="patient-cargivers-table-header-email">
+              {t("email")}
+            </TableCell>
+            <TableCell id="patient-caregivers-table-header-actions" />
           </TableRow>
         </TableHead>
         <TableBody>{tableRows}</TableBody>