+++ packages/yourloops/test/pages/profile/profile.test.tsx
@@ -44,9 +44,9 @@ import { Preferences, Profile, Settings } from "../../../models/shoreline";
 
 function testProfile(): void {
   let container: HTMLElement | null = null;
-  let updatePreferences: sinon.SinonStub<[Preferences,boolean|undefined], Promise<Preferences>>;
+  let updatePreferences: sinon.SinonStub<[Preferences, boolean | undefined], Promise<Preferences>>;
   let updateProfile: sinon.SinonStub<[Profile, boolean | undefined], Promise<Profile>>;
-  let updateSettings: sinon.SinonStub<[Settings,boolean|undefined], Promise<Settings>>;
+  let updateSettings: sinon.SinonStub<[Settings, boolean | undefined], Promise<Settings>>;
   const defaultUrl = "/professional/patients";
 
   async function mountProfilePage(session: Session): Promise<void> {
@@ -120,6 +120,7 @@ function testProfile(): void {
 
   it("should display pro sante connect button if user is a french hcp and his account is not certified", async () => {
     const session = loggedInUsers.hcpSession;
+    session.user.frProId = undefined;
     await mountProfilePage(session);
     const proSanteConnectButton = container.querySelector("#pro-sante-connect-button");
     expect(proSanteConnectButton).to.be.not.null;
@@ -127,10 +128,24 @@ function testProfile(): void {
 
   it("should display eCPS number if user is a french hcp and his account is certified", async () => {
     const session = loggedInUsers.hcpSession;
-    session.user.frProId = "ANS12345789";
     await mountProfilePage(session);
     const textField = container.querySelector("#professional-account-number-text-field");
     expect(textField).to.be.not.null;
+  });
+
+  it("should display certified icon if user is a french hcp and his account is certified", async () => {
+    const session = loggedInUsers.hcpSession;
+    await mountProfilePage(session);
+    const certifiedIcon = container.querySelector(`#certified-professional-icon-${session.user.userid}`);
+    expect(certifiedIcon).to.be.not.null;
+  });
+
+  it("should not display certified icon if user is a french hcp and his account is not certified", async () => {
+    const session = loggedInUsers.hcpSession;
+    session.user.frProId = undefined;
+    await mountProfilePage(session);
+    const certifiedIcon = container.querySelector(`#certified-professional-icon-${session.user.userid}`);
+    expect(certifiedIcon).to.be.null;
   });
 
   it("should update profile when saving after changing firstname", async () => {